Backend Server/OTA/ota_server/requirements.txt
flask==3.0.2
pyyaml==6.0.1
python-dotenv==1.0.1
flask-limiter==3.5.0
flask-cors==4.0.0
cryptography==42.0.5
requests==2.31.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0 